Extract ExternalChangesetTitle to shared component file, refactor ExternalChangesetInfoCell to employ it
