+++ client/web/src/enterprise/batches/detail/changesets/ExternalChangesetInfoCell.tsx
@@ -1,9 +1,7 @@
 import classNames from 'classnames'
-import ExternalLinkIcon from 'mdi-react/ExternalLinkIcon'
 import React from 'react'
 
 import { Link } from '@sourcegraph/shared/src/components/Link'
-import { LinkOrSpan } from '@sourcegraph/shared/src/components/LinkOrSpan'
 
 import {
     ExternalChangesetFields,
@@ -13,6 +11,7 @@ import {
 
 import { ChangesetLabel } from './ChangesetLabel'
 import { ChangesetLastSynced } from './ChangesetLastSynced'
+import { ExternalChangesetTitle } from './ExternalChangesetTitle'
 
 export interface ExternalChangesetInfoCellProps {
     node: ExternalChangesetFields
@@ -24,67 +23,50 @@ export const ExternalChangesetInfoCell: React.FunctionComponent<ExternalChangese
     node,
     viewerCanAdminister,
     className,
-}) => (
-    <div className={classNames('d-flex flex-column', className)}>
-        <div className="m-0">
-            <h3 className="m-0 d-block d-md-inline">
-                <LinkOrSpan
-                    to={node.externalURL?.url ?? undefined}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    className="mr-2"
-                >
-                    {(isImporting(node) || importingFailed(node)) && (
-                        <>
-                            Importing changeset
-                            {node.externalID && <> #{node.externalID} </>}
-                        </>
-                    )}
-                    {!isImporting(node) && !importingFailed(node) && (
-                        <>
-                            {node.title}
-                            {node.externalID && <> (#{node.externalID}) </>}
-                        </>
-                    )}
-                    {node.externalURL?.url && (
-                        <>
-                            {' '}
-                            <ExternalLinkIcon size="1rem" />
-                        </>
-                    )}
-                </LinkOrSpan>
-            </h3>
-            {node.labels.length > 0 && (
-                <span className="d-block d-md-inline-block">
-                    {node.labels.map(label => (
-                        <ChangesetLabel label={label} key={label.text} />
-                    ))}
-                </span>
-            )}
-        </div>
-        <div>
-            <span className="mr-2 d-block d-mdinline-block">
-                <Link to={node.repository.url} target="_blank" rel="noopener noreferrer">
-                    {node.repository.name}
-                </Link>{' '}
-                {hasHeadReference(node) && (
-                    <div className="d-block d-sm-inline-block">
-                        <span className="badge badge-secondary text-monospace">{headReference(node)}</span>
-                    </div>
+}) => {
+    const changesetTitle =
+        isImporting(node) || importingFailed(node)
+            ? `Importing changeset ${node.externalID ? `#${node.externalID}` : ''}`
+            : `${node.title} ${node.externalID ? `(#${node.externalID})` : ''}`
+
+    return (
+        <div className={classNames('d-flex flex-column', className)}>
+            <div className="m-0">
+                <ExternalChangesetTitle className="m-0 d-block d-md-inline" externalURL={node.externalURL}>
+                    {changesetTitle}
+                </ExternalChangesetTitle>
+                {node.labels.length > 0 && (
+                    <span className="d-block d-md-inline-block">
+                        {node.labels.map(label => (
+                            <ChangesetLabel label={label} key={label.text} />
+                        ))}
+                    </span>
                 )}
-            </span>
-            {![
-                ChangesetState.FAILED,
-                ChangesetState.PROCESSING,
-                ChangesetState.RETRYING,
-                ChangesetState.UNPUBLISHED,
-                ChangesetState.SCHEDULED,
-            ].includes(node.state) && (
-                <ChangesetLastSynced changeset={node} viewerCanAdminister={viewerCanAdminister} />
-            )}
+            </div>
+            <div>
+                <span className="mr-2 d-block d-mdinline-block">
+                    <Link to={node.repository.url} target="_blank" rel="noopener noreferrer">
+                        {node.repository.name}
+                    </Link>{' '}
+                    {hasHeadReference(node) && (
+                        <div className="d-block d-sm-inline-block">
+                            <span className="badge badge-secondary text-monospace">{headReference(node)}</span>
+                        </div>
+                    )}
+                </span>
+                {![
+                    ChangesetState.FAILED,
+                    ChangesetState.PROCESSING,
+                    ChangesetState.RETRYING,
+                    ChangesetState.UNPUBLISHED,
+                    ChangesetState.SCHEDULED,
+                ].includes(node.state) && (
+                    <ChangesetLastSynced changeset={node} viewerCanAdminister={viewerCanAdminister} />
+                )}
+            </div>
         </div>
-    </div>
-)
+    )
+}
 
 function isImporting(node: ExternalChangesetFields): boolean {
     return node.state === ChangesetState.PROCESSING && !hasHeadReference(node)
